Make MediaButtonReceiver MediaBrowserServiceCompat aware
While maintaining the ability to have media buttons routed
to a particular service by adding the MEDIA_BUTTON intent
filter, also look for and use any available media browser
services as possible targets to forward media button events.

This saves a step in updating the manifest and fixes the
common case: using a MediaBrowserService(Compat) as the
Service owning the MediaSessionCompat.

BUG: 22917960
Change-Id: Ic0db49957a8bf23d378cafbdd588ba1dd39b2da5