Fix AppBarLayout layout issues
Cropped up after we moved to laying out the scrolling view
in the offset position. The problem occurs when the
AppBarLayout is scrolled off screen by any amount, and
then a layout occurs. The scrolling view is then laid out
attached to the bottom of it.

The old scrolling logic didn't account for this and expects
the scrolling view to be laid out in the same base position
every time.

Since we now lay the scrolling view correctly in the
first place, there is now no need to use ViewOffsetHelper
and can just use offsetTopAndBottom to offset while scrolling.

Change-Id: If9b4cb9ebae49235ffbc152a5f7af87843b459e7