Add a way to put more than one custom action button to customtabs
This adds new APIs to CustomTabsIntent/CustomTabsSession to add/update
more than one button to the UI with assigned ID. To keep backwards
compatibility with the previous APIs, the one main button is always
given a constant ID and new ones are assigned by the client. This uses
the already present updateVisuals aidl call to send the messages
to the browser side.

BUG=26470162

Change-Id: Ic6668093f494796a1e91508b0a29ef2b5760875a