More tests for DrawerLayout.
Add tests that cover the internal DrawerLayout.isDrawerView via a variety
of API calls - opening / closing a drawer, checking open state, setting
lock mode and querying the lock mode.

Also add the same isDrawerView check to getDrawerLockMode as we have in
isDrawerOpen, isDrawerVisible and various setters that get View as the
parameter.

Finally, annotate all the existing DrawerLayout tests so that they can properly
run on Sponge and simplify the inner layout of the drawer content.

Bug: 26642831
Change-Id: Ia62b415eda39a2a6560b15cc0a55e6d227261a6c