Fix onStart/EndAction() for ViewPropertyAnimatorCompat
These methods are meant to be single-action only (removed when they
are activated). That's the way they work on the framework; it was an oversight
to have them be persistent in the ICS version of the support library.

Issue #25764942  ViewPropertyAnimatorCompatICS end action is not cleared

Change-Id: I0825330c76cecf63960d79f74116acd99b874830
(cherry picked from commit fdd92b54cc21d9927ee89cf675cf4c90c306f24a)