Adjust ViewDragHelper to better support watches.
ViewDragHelper hard-codes the edge size to 20dp for all devices.
On watches, this is too small. Watches should define the edge size
as a percentage of the screen height. This introduces a new method
ViewDragHelper.refreshEdgeSize() to calculate the edge size. For
phones, this will continue being 20dp, but for watches edge size
will be calculated.

Change-Id: I2d8521fab702d447b63b8508634230a5699e1a43