First pass for testing anchoring of views in CoordinatorLayout.
One test checks that snackbar shown in coordinator layout with
a FAB pushes the FAB up as a built-in behavior. The second test
checks that snackbar showin in coordinator layout results in
custom view's behavior (registed in XML as class name) to be invoked.

The setup is dynamic, allowing using multiple "content" layouts in
the same activity so that we don't need to have multiple Activity
classes cluttering the package.

This CL also has a proposed way to "wait" until Snackbar has been
fully shown with custom callback and Espresso's idling resources.

Change-Id: Idf0ab8cdc7e2454bbcddb36c94d832d921c56da6