Fix crash: Dont fetch items in onMeasure when HasFixedSize is true
With new RV change, HasFixedSize=true hehavior has been defined as:
RV size is not determined by child size.  So when HasFixedSize is
true, the items passed to onMeasure() are no longer valid and should
not be used.  Though HasFixedSize=true is generally not allowed when
WRAP_CONTENT is used,  add extra protection.

Bug 27202689

Change-Id: Ic1e30674588227988621f7802692a1f12eeac3d7