Handle gracefuly if changed items disappears
This CL fixes a bug where if an Item is changed and also
disappears in the same layout pass, RecyclerView would call
both animation APIs which is a problem because it is designed
to call only one animateXXX callback per ViewHolder.

With this CL, it only calls animateChange but takes care
of hiding the disappearing views.

Bug: 27137909
Change-Id: I28e7a3c24e9cd0c8bf6790070b382830014e8987