Stop PFD finalizer from closing developer sockets.
When creating a PFD, it takes ownership of the underlying FD and
its finalizer will try closing the FD once we return from our tagging
methods.  To prevent the unwanted close, we detach the FD, but we
need to wait until after the tagging operation has taken place, since
it needs a real FileDescriptor object.

Bug: 27274605
Change-Id: I6e5eb031bb354d4f9fe0f2c523ee817ce687c437