Protect FragmentManager against View.NO_ID container ids
Not all code paths for FragmentManager were checking that the
container view had a valid view id. As we can't correctly restore a
fragment with a container without one, throw a more descriptive
exception earlier.

Update the FragmentPagerAdapter implementations to be even more
descriptive if the host ViewPager does not have a view id and we would
enter this case.

Bug 27290033

Change-Id: I3e349ceaa82fbbd1f257554d9670e4c21f49ff4e