Get direct child in focus failure
This CL fixes a bug in SGLM where if a it is called with a sub child
of an item view, it tries to get the LayoutParams and fails to cast it.
This CL makes sure SGLM uses the direct child while doing these controls.
Other layout managers already handle this.

Bug: 27153378
Change-Id: If7ba2012812c40b8898d1876e6cf68a5f818736c