Leanback Support Library: Fixed issue with GuidedStepFragment actions disappearing after an action is collapsed
The recycler view, by default, prunes children of mActionsGridView
when they were pushed off of the screen after expanding actions.
Fixed that by disabling the pruning behavior before an action
is expanded (after the first click on an action),
and enabling it back after the action is collapsed
(after the second click on that same action).

Bug: 27482284
Change-Id: Ibf0ba8c3bae271059e6bea0202fefe9b16acc867