Handle hint in unspecified specs
On M+, the parent can pass a hint size with unspecified specs.
Previously, our LayoutManagers ignored this because it might
be garbage on older platforms.

With this CL, RV ensures that it is 0 on older platforms and
the value parent provided in newer ones. This way, RV can
know when to stop based on the hint but still ask for what
it needs (so that it is not treated as AT_MOST).

Bug: 27608240
Change-Id: I2d452731c7251b0ee1f35b9a0694c78381cdf39f