Back out TintResources changes DO NOT MERGE ANYWHERE
They were causing memory and configuration issues
therefore we need to back it out. This will revert
the functionality of being able to reference vector
drawables from within resources, but it's a necessary
cost.

We still need TintResources, but it's now back to being
a wrapper object which does not cause any issues.

BUG: 27877272
BUG: 27808232

Change-Id: Iad401718e51da15fddbd7ff88e41743dca40ad04