Adds a launching activity flag to WearableExtender in NotificationCompat
This adds a flag to WearableExtender (and Action.WearableExtender)
indicating that the attached ContentIntent will launch an activity
on Wear, which lets the platform know that it can (and should)
generate a sleek transition for the launching of said intent (as
opposed to the intent causing change in the background).

This also updates the api. Hooray.

Bug:28040246
Change-Id: I5ca929a9ed0be90726a8e0f333f68ec7fba6cac2