Add vectors within resources back behind an API
This adds back in the functionality removed in
Iad401718e51da15fddbd7ff88e41743dca40ad04, but
behind an API, defaulting to disabled.

Removing functionality wasn't good, but we needed
to do it. This enables developers to use the functionality
again, but having to explicitly opt in with the
associated caveats.

BUG: 28186451

Change-Id: Ic99b6319fdd58a5a6ece19b2fbf88c148bfd0618