Invalidate anchor info only after a full layout
This CL fixes a bug in built in layout managers. Due to
intermediate measure steps with different sizes, the layout
manager may lose its anchor position since it always tries
to sync with the previous measurement which might be much
greater or smaller than the final size.
The same problem was happening with scrollToPosition calls
which was previously fixed with the onLayoutCompleted API
and this CL fixes relayout cases using the same API.

Bug: 27787175
Change-Id: I1eebb652a402022c99c32fad57e0085a461bf117