Remove BaseFragmentActivityApi24 and move multi-window methods
On L and above the ART compile-time verifier can't handle calls to
superclass methods that didn't exist yet. This causes the verification
to be deferred to runtime, which equates to a very large startup time
regression for activities extending FragmentActivity or
AppCompatActivity.

We don't want apps to need to call super for these methods anyway; a
corresponding change in frameworks/base moves the additional internal
Activity dispatch logic into private methods that invoke the public
ones. Apps extending FragmentActivity or AppCompatActivity still need
to call super and the FragmentActivity methods are annotated as such,
but now they won't pay the additional verifier costs at the support
lib/framework boundary.

Bug 28667205

Change-Id: Id5b12322ec1b5544f6bfa6e793560df8058b4314