Leanback playback: Added support to flip between different views for each media item row depending on the playback state
The first view of each media item row can now flip between different
views. A default layout is provided by the AbstractMediaItemPresenter or
the user can set a theme attribute that defines a custom layout for
different playback states.
In the default layout, a 3 peak meters animation is provided when a
media item is playing.

Change-Id: Ibcc8d995a9d46cf089afc998980c815948609371