Automatic peekHeight for bottom sheets
Apps can specify app:behavior_peekHeight="auto" to let the
BottomSheetBehavior automatically calculate the 16:9 ratio keyline and
use it as the peek height while keeping it at least as tall as 64dp.

This is now the default parameter of BottomSheetBehavior and
BottomSheetDialog. The old default was 0, so this change will not affect
existing apps which should already have some value specified for the
peek height.

Bug: 27260758
Change-Id: Ibbf725aa89352512f4b54fa9fe297c1675f76939