Fix CTL scrim not working with activity transitions
Seems to be an issue with how transitions draws
its views, and how CTL draws the scrim. Fixed by
moving the scrim draw command to the preceding view
before the Toolbar.

Also fixed an issue where the title would be drawn
in the wrong position when we have window insets.

BUG: 29748380

Change-Id: I83b9a7ae894cd36ab9ca72cd1fe5d111aa6f2b07