Expose external app files/cache dir from FileProvider.
This adds support for an application's external files directory and cache directory
into FileProvider. It also corrects the incorrectly-documented <external-path> support
to correctly reference that it is the external storage root.

With this the FileProvider tests were also ported to run on Gradle, and updated with
tests for the new tags.

Resolves:
- http://b.android.com/67171
- http://b.android.com/184603
- http://b.android.com/211791

Change-Id: I04d514ebe99ab66ae59a9489f5fd3d7769cb79fb