Add methods to opt out from launching custom tabs
Added a private extra to CustomTabsIntent that when added with NEW_TASK
will override any customization and with or without any other extras avoid
launching a custom tab. Added a convenience call which add this and
NEW_TASK to a given Intent. When using NEW_TASK for the purposes of
getting a browser UI becomes the norm, the added extra can be safely
deprecated. Also added a method for browsers to check these flags and
extras

BUG=30572193

Change-Id: I138d68d00bed91b08d2fb99ac738bf38a44a39a2